v0.2.13-feat: Refactor eAffine class into its own module and update imports
- Moved eAffine class from normalization.py to affine.py
- Updated version number in pyproject.toml to 0.2.13
