+++ symm_learning/nn/__init__.py
@@ -1,8 +1,9 @@
 from .activations import Mish  # noqa: D104
+from .affine import eAffine
 from .conv import GSpace1D, eConv1D, eConvTranspose1D
 from .disentangled import Change2DisentangledBasis
 from .distributions import EquivMultivariateNormal, _EquivMultivariateNormal
-from .normalization import eAffine, eBatchNorm1d
+from .normalization import eBatchNorm1d
 from .pooling import IrrepSubspaceNormPooling
 
 __all__ = [
